Rotate the whole thing so that Z points north instead of Y
This is all matter of convention, but references such as
https://www.academia.edu/8612052/ORBITAL_MECHANICS_FOR_ENGINEERING_STUDENTS
or
http://www.braeunig.us/space/orbmech.htm
rarely assume Y the normal vector of orbit plane.
Usually, XY is the orbit plane and Z is the normal vector.

On the other hand, OpenGL coordinate system uses XY as the
screen coordinates, which is okay with 2-D plot of orbits but
do not cope well with 3-D view with OrbitControls.