Move all objects in scene relative to centered object for removing floating errors
When we zoom to a planet so close that we can see the shape of it, the
floating errors become prominent that the planet looks trembling.
So we subtract centered position from all other objects in the scene
to make the center look like focused.
This way, the offset won't be multiplied with less precise transformation
matrices.