Add ModulatedIcosahedronGeometry that is copied from THREE.js code
Because THREE.js later than 0.125.0 does not allow changing vertex positions
in IcosahedronGeometry (or any geometry in general), we need to derive and
create our own copy of IcosahedronGeometry (and in turn PolyhedronGeometry
that it inherits from).

In order to make a bumpy rocky geometry for the asteroids, we need to recreate
IcosahedronGeometry compatible thing on our own.