Notify the client when a new rocket was added
Since clientUpdate message doesn't have enough information to
reconstruct the original object, we need a new message to notify
the client that new object was added with complete information.

It is similar to "load" api, but only about once object.
It has a little inconvenience that parent's id won't work.